Fix ext-partition + motion-var
When calculating the above overlapped prediction, we adjust
xd->mb_to_bottom_edge to have the correct value for the prediction
we're about to make. This prediction is half the height of the
current block - *unless* the block is 128 pixels high, in which
case the prediction is only 32 pixels high.

This patch implements the special case for 128-pixel high blocks.

(all of this applies analogously for left predictions too)

Change-Id: I2780e8d4a6f8a9f82f540f2c2dfde691d07298e2