Update patch set 2
Patch Set 2:

Hi David,

Yes it's a good point, the tile group handling needs log2_tile_rows/cols set up to work properly. The first patch was only intended to support uniform tiles. There is also a fix needed to tile group header writing in the encoder for when number of tiles is not a power of two (see change to bitstream.c in this patch).

The status of this patch is that I think it is complete and I would like to submit. Just waiting to see if there are any review comments from people with related experiments.

Yunqing or Fangwen - are you OK with the patch? If so please can you +2 on code review, or let me know of comments if there is an issue.

Thanks,
Dominic

Patch-set: 2
Reviewer: Gerrit User 7984 <7984@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>