Update patch set 3
Patch Set 3:

Thanks for the review, and I will submit.

Max tile does actually allow for heights that are greater than 2304, just that the area is limited to 4096x2304 (so you are permitted to have a narrow high tile).

Explicit tiles with different sizes are supported in the syntax and decoder, but there is currently no encoder API. I'm not sure how necessary this is - do you have any thoughts on encoder API?

Patch-set: 3