Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Thanks for the review, and I will submit.
> 
> Max tile does actually allow for heights that are greater than 2304, just that the area is limited to 4096x2304 (so you are permitted to have a narrow high tile).
> 
> Explicit tiles with different sizes are supported in the syntax and decoder, but there is currently no encoder API. I'm not sure how necessary this is - do you have any thoughts on encoder API?

In explicit tiles, each tile's sizes are transmitted. If there is no encoder API, how it can be used? At least it should allow entering a list of tiles with different sizes. Maybe put some restrictions on allowed tile sizes in a frame. What are the possible use cases for this?

Patch-set: 3