Update patch set 3
Patch Set 3:

> In explicit tiles, each tile's sizes are transmitted. If there is no encoder API, how it can be used? At least it should allow entering a list of tiles with different sizes. Maybe put some restrictions on allowed tile sizes in a frame. What are the possible use cases for this?

I think the use cases mentioned in the working groups are for load balancing of number of bits and for region of interest coding. I think both Fangwen and Thomas Davies have mentioned such use cases.

The purpose of the max_tile additions is to ensure that explicit tiles are forced by syntax to obey the max tile limits proposed to the group. Use by an encoder for the above use cases is up to the specific encoder.

However, I agree that having a list of tile sizes as input is useful for testing and will look further at this. For aomenc I did not notice any easy way of specifying a list so started with the individual setting width/height as discussed previously. I will look at the possibility of a patch for adding a list format as an input.

Patch-set: 3