Adjust QP for AOM_Q (constantq) rate control mode
For the AOM_Q RC mode, obtain the gf_active_quality first for a GF
group and adjust the cq_level for different frame update types
within the GF group.

The cq_level for ALTREF and ALTREF2 frames had been adjusted
previously. This patch adds the tuning on the cq_level for BWDREF and
the bi-predictive frames in between.

Change-Id: I9cba639cd319f8ee0a5821df4ddfdb4e7bf34c1e