fix one bug in the decoder header initization
1. seg->temporal_update flag is only read when seg->update_map flag is
set to 1 in the bit stream. in case update_map flag is different from frame
to frame, the temporal_update flag will not be reset to 0 when
update_map flag changed from 1 to 0. i.e, the previous temporal_update
flag will be used in the later process. in the read_inter_segment_id
function, temporal_update flag is used to read the segment id. it will
cause bit stream parsing error.

Change-Id: I663975bc84bbed7409d4f5dd1c1005480b2fbcf8