Change loopfilter logic: consider ref frame and MV
Previously, if two blocks on either sides of filtering edge are skipped,
the edge will not be deblocking filtered. But there're cases where these
two blocks are skipped but are not motion compensated from the same
reference frames and motion vectors.

Add the constrain of same ref frame and MVs to enforce more deblocking
filtering.

Performance is identical on AWCY.

Change-Id: Iabb88433a471694874faef9373ec1aca4d6d774a