Add decoder support to ext_skip
At the frame header level, add the checking on whether an inter-coded
frame allows the use of 'skip_mode'. At the block level, add the
decoding of 'skip_mode' for all the blocks if 'skip_mode' is allowed
for the current frame. If 'skip_mode' is set, no further bit needs to
read out, and the block is set as:
- Inter-coded in compound mode;
- With the nearest forward and backward reference frames;
- In NEAREST_NEARESTMV mode;
- In SIMPLE_TRANSLATION motion mode;
- In COMPOUND_AVERAGE compound type; and
- Skip set for residual.

Change-Id: I50297f63a81135109332c4a51b516bffd7726777