Add experiment ext_skip
This coding tool is to introduce a new prediction mode for the
bi-predictive frames that have a forward referernce within 2 frames
away (distance denoted as 'fwd_delta'), and a backward reference,
within (3-fwd_delta) frames away.

If this prediction mode, namely 'ext_skip' is set, it will be coded
using compound prediction with the most recent forward and backward
reference frames as its reference pair, NEARESTMV as its motion mode,
and the skip flag is set for the residue.

Change-Id: I826034ccf1a956f4b350f0bc2e2dca8ea71b5197