Deduplicate use_reference_buffer/frame_id_numbers_present_flag
These two flags serve essentially the same purpose, so we can
simplify down to just having one flag. Since the
frame_id_numbers_present_flag is older and listed in the
reference-buffer design doc, this is the one we'll keep.

Note that, previously, frame_id_numbers_present_flag was only
sent when CONFIG_OBU was enabled - otherwise it was always set
to 1. On the other hand, use_reference_buffer was always signalled
so that allowed us to disable frame ids even without CONFIG_OBU.

In order to keep the ability to disable frame IDs without
CONFIG_OBU, we send a "real" sequence header (ie, with the same
data as in CONFIG_OBU) in-line within the uncompressed header.
This is sent every keyframe and every intra-only frame, just
as the use_reference_buffer flag was before this patch.

BUG=aomedia:853

Change-Id: I4bb6c7a222aeddfb6d3737a39d10782d32ee2eee