Update patch set 3
Patch Set 3:

> You are treating the symptom not solving the problem.
100% agreed.
Please note that this changelist is not a fix.
Please also note that the assertions are there to inform the developers, not to protect the users.

The underlying problem is that the implementation of the new transforms (the ones that are enabled by txmg) assume "typedef int32_t tran_low_t". This could be fixed by writing an alternative implementation of those transforms.

> This could be solved at compile time, by not letting people --disable-highbitdepth.

Indeed, and this is what I did, but only when this is actually required, i.e when txmg is enabled. This is where the users are protected.

In one word: in the current state, I think it's impossible to trigger these assertions with an unmodified working copy.

Does it make sense?

Patch-set: 3