Don't generate TEXT relocations in asm code with CONFIG_PIC
Assembly code in aom_dsp is assembled without -fPIC (I'm not sure why,
but assume this is for some tooling reason). This means that the code
can't use its handy snippets of .rodata when building a library.

The filter_block1d16_ macro in aom_subpixel_8t_ssse3.asm used some
rodata even when CONFIG_PIC was enabled. This patch teaches it not to
and also hides the .rodata declarations in this case (so we get
compile-time errors, rather than a cryptic message from Jenkins if we
forget!)

Change-Id: Ibb16c76fd593b35dc37ce0425d288e6a405e2e84