Rework key frame intra mode context model
Reduce the context model size for key frame modes from 30240 bits
to 4500 bits, i.e., less than 1/6 of the original context model.
The coding performance loss on key frame is 0.14% for lowres and
noise level difference for video sequence. The loss on key frame
for midres is 0.05% and noise level for whole video.

Change-Id: I9e36825e5c5ee6ba35038c3ca349ad1ad3429910