Fix get_plane_block_size => max_block_(wide|high)
Before this patch if you had a block size, bsize, and you wanted to
get the width of that block in the U plane, you'd do something like:

  plane_bsize = get_plane_block_size(bsize, pd);
  blocks_wide = max_block_wide(xd, plane_bsize, plane);

That is ok, but doesn't work for small rectangular blocks. For
example, if bsize is 8x4 and we're only subsampling in the Y direction
then get_plane_block_size would have to somehow return
BLOCK_8X2 (which doesn't have an enum value).

Before this patch, the code above would then do a global buffer
overrun because plane_bsize gets set to BLOCK_INVALID (the only thing
it could reasonably be) and then max_block_wide uses that as an array
index.

This patch changes max_block_wide and max_block_high to take the block
size before subsampling. They also now understand CHROMA_SUB8X8 so
will round sizes up to 4 if necessary.

Going through the call sites, there's lots of duplicated code for
computing plane_bsize that can just be deleted.

BUG=aomedia:875

Change-Id: I7d8849701550f85c521c0a03c4182357998f2bc3