Use min tx type for vartx transform type search
This addresses an inconsistency between the set used
to send the tx_type in the bitstream and the set used
for the tx_type search. Previously, the set used to
read/write the tx_type was based on the smallest tx_size
in the vartx partitioning, but the search uses a set
based on the largest possible tx_size. This patch
changes the tx_type search to use the transform type
set associated with the smallest possible tx_size to
make the search more consistent with the bitstream syntax.

This patch also adds assertions to all exposed transform
functions to ensure that no illegal transform type/size
combinations are occuring.

This currently gets about a 0.1% drop in performance,
presumably because it increases the number of small tx sizes.
I am currently trying an alternate version to see if it
can reduce the drop.

Change-Id: I10d46877a1d3e37dd7e5e609770f7a1c793241a1