Update patch set 6
Patch Set 6:

It is okay for now, but I believe the current design will not work if the tilewidth is not a multiple of the restoration tile size.
This can happen if cm->sb_size is smaller than the restoration tile size. We do need to fix this case.

Patch-set: 6