OBU: Check OBU sizes
Add a proper error message for the case where the signalled
size of an OBU is different to (obu_header_size + obu_payload_size).

This can happen in various ways - for example, if a temporal
delimiter is sent (which always has header+payload size of 1
or 2 bytes), but is signalled with a larger OBU size.

Change-Id: Ia92b589bb4e869e43e595b79e0e657380acb499d