[non-normative, obu] Check OBU sizes
Add proper error detection for the case where the signalled
size of an OBU differs from the actual decoded size

This requires explicitly calculating the expected payload size
(based on the OBU size and OBU header size, depending on experiments),
so we can simplify the code to handle metadata/padding OBUs slightly.

Note: This patch preserves the current decoder behaviour, which
is to error out if the expected payload size and the true payload
size differ. It just promotes that from a desync issue to an
explicit error message.

But this patch is written to make it trivial to allow the
signalled size to be larger than the true size (treating the
extra bytes as skippable padding), if we wanted to make that
normative change later.

Change-Id: Ia92b589bb4e869e43e595b79e0e657380acb499d