Update patch set 4
Patch Set 4:

I've brought this patch up to date and also done a little bit of code cleanup. As per Peter's recent comment on the attached bug report, we now feel that this is the best option for how to select cm->prev_frame.

A quick test on our side suggests that the change is ~neutral with respect to BDRATE, but I'll set up an AWCY run to check.

Patch-set: 4
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>