Avoid Visual Studio compile error in loopfilter
If you have a structure, foo_t, with an alignment request then Visual
Studio won't allow you to declare a function

  void use_foo(foo_t x);

The reasoning is that x might be passed on the stack, and their ABI
doesn't allow them to guarantee that x is aligned appropriately. More
strangely, this isn't allowed either:

 void use_some_foos(foo_t x[10]);

This is functionally equivalent to:

 void use_windows_foos(foo_t *x);

(except that you can't tell how long the array should be from the
function signature).

Since Visual Studio is supposed to allow the latter form, use that
instead.

Change-Id: Icd449fc1058606fa7e48a6f791091bbb42a73b2c