Don't choose partitions that aren't allowed for current subsampling
Before this patch, there was nothing to stop the encoder choosing a
partition into 4x16 blocks (say) when only subsampling in the x
direction. This would give a 2x16 block in the chroma plane and fall
foul of bitstream conformance.

I found this while trying to fix aomedia:897, but that turns out to be
an unrelated problem.

Change-Id: I35f175236188fb7c3c182ef1f5a8a12f8f6d761f