Update patch set 1
Patch Set 1: Code-Review-1

Marking as WIP and code review = -1 for the moment, as I still need to implement the SEG_LVL_ZEROMV side of this patch. That should be ready in the next day or so.

Patch-set: 1
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1
Work-in-progress: false