Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> (1 comment)

If mi_col = 1, we adjust to use prev_mv corresponding to mi_col = 0, in this case, if blk_col = 1, we should use prev_mv corresponding to mi_col = 2. That is why it is rounded up.

Actually, since mi_step in the caller function is always a even number, so here it doesn't matter. Removing the rounding part makes it simple. I will modify it to blk_col >> 1.

Patch-set: 12