[CFL] Sub8x8 Validation Code Rewrite
Sub8x8 Validation code is changed to be more robust. The scope of the
validation is narrowed to validating that all of the required content in
the storage buffer was stored between CfL predictions. The early
termination used in the current mode decision code does not allow to
validate more than that.

This change does not change encoder output

Change-Id: I7f1ed84da5037dcfaaf5da9cf33b4b8d664d2352