Update patch set 2
Patch Set 2:

> I couldn't look into this issue yet. But if you use Perl script,
 > your avx2 function should not contain a C version to simulate AVX2
 > intrinsics. Avx2 function (*_avx2()) content should map to avx2
 > instruction set directly. Please correct me if I'm wrong.

True, the C code is only meant as a fallback which shouldn't happen at all.  It's mainly there for debug and porting convenience.  Perhaps it should issue a #warning (but the C code _is_ intentionally used in the unit tests).  The purpose of the v64/v128/v256 types and operations on them is to be able to write the code once and still get specialisations for arm, sse2, ssse3, etc.  Only when the architecture that you compile on is unsupported by the v64/v128/v256 types, there should be a fallback to C code, just to get something functionally correct.

I added unit tests to catch unintended issues like this (if indeed slow C code for some reason get run), but they caused occasional issues when run on loaded systems, so they had to be disabled.

Patch-set: 2
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>