Update patch set 1
Patch Set 1:

I'm not sure that this is the right solution to the linked bug report - the encoder is expected to check the validity of the warped motion model, and it's only supposed to signal WARPED_CAUSAL if that will result in a valid warp model.

It's very likely that there's an underlying bug on the encoder side, which needs to be fixed instead. This patch might hide the bug for the moment, but it's likely to reappear as encode/decode mismatches later.

Patch-set: 1
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>