Update patch set 2
Patch Set 2:

(1 comment)

> [...]
> 
> Things would get a lot easier if we would enlarge the set of valid bitstreams (and thus, shift the issue from a compliance issue to a quality issue).
> Encoder vendors still have to check their output, but no downstream decoder will crash if an unexpected matrix is transmitted.

That makes sense, and I agree that it makes things easier if there are fewer bitstream conformance requirements.

In that case, I'm happy with this patch as long as Sarah/Yue/Deb are. You might also want to make a similar change to read_global_motion().

Patch-set: 2
CC: Gerrit User 5188 <5188@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>