Update patch set 2
Patch Set 2:

Hi there,

David and I have just been looking at this and think there's a slight problem. If delta_frame_id_length is coded as 0b1111 = 15, you get a delta_frame_id_length of 17. If the next 4-bit literal is also 0b1111, you end up with a frame_id_length of 33. That doesn't work with the current code, because when we actually read frame_id_length, we use aom_rb_read_literal with return type int (a signed 32 bit integer on most platforms).

If I've got this right, we used to have delta_frame_id_length in the range [2, 17] and frame_id_length in the range [7, 22]. With the current patch, this changes to [2, 17] and [3, 33] respectively. Could we read a 3-bit literal when calculating frame_id_length instead? Then we'd have ranges [2, 17] and [3, 25] which is still more generous than it was before (and uses one less bit to encode) but fits in a standard integer type.

Rupert

Patch-set: 2
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Reviewer: Gerrit User 7560 <7560@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>