Add motion selection to ext_skip
A new block mode, referred to as skip_mode is added. If a block is
coded as skip_mode, it will be inter-coded, with its references and
motion vectors derived from its neighboring blocks, spatially or
temporally (if allowed), with zero-residue. Otherwise, the block
can be coded in the current intra or inter mode.

Skip_mode is now only allowed for bi-predictive frames, under the
constraint that its most recent forward reference and most recent
backward reference are no more than 3 frames away.

The reference frames and motion vectors are derived from the mvs of
the spatial and temporal neighboring blocks in a certain order. Both
single and compound modes are considered.

The computational load on skip_mode evaluation at the encoder should
be kept minimum. No transform size / type evaluations are needed.

Change-Id: I5aef0159c7d5ecd64258510835903375d6c536d6