Add encoder support to ext_skip
A new block mode, referred to as skip_mode is added. If a block is
coded as skip_mode, it will be inter-coded, with its references and
motion vectors derived from its neighboring blocks, spatially or
temporally (if allowed), with zero-residue. Otherwise, the block
can be coded in the current intra or inter mode.

Skip_mode is now only allowed for bi-predictive frames, under the
constraint that its most recent forward reference and most recent
backward reference are no more than 3 frames away.

The computational load on skip_mode evaluation at the encoder should
be kept minimum. No transform size / type evaluations are needed. The
reference frames and motion vectors are derived, and the RD cost of
skip_mode is finally compared against the best RD mode from the
evaluation loop of MAX_MODES without the consideration of skip_mode.

Regardless of the status of mbmi->skip_mode, keep the bitstream
handling on cm->delta_q_present_flag;

Change-Id: I5aef0159c7d5ecd64258510835903375d6c536d6