Update patch set 7
Patch Set 7:

I was just looking over this patch again, and thought: What if, instead of setting up a default warp model when the signalled one is invalid, we instead just reset the motion mode to SIMPLE_TRANSLATION?

The main advantage of this approach is that it preserves the block's motion vector, whereas the current patch effectively ignores it and applies a motion of (0, 0) instead. You could of course build a model which has the right translational part and go through the warp filter path, but it's a lot more work.

Patch-set: 7