Remove partial_frame support from loop restoration
This flag comes from the loop filter's speed features and (I think)
tells the encoder to make decisions about the filter by looking at a
narrow strip in the middle of the frame.

That's reasonable enough, but doesn't make any sense for loop
restoration, where we were calling av1_loop_restoration_frame from
pickrst.c in order to calculate what restoration parameters to use for
a given restoration unit (which might not be in the narrow strip in
the middle!)

As it turns out, the LPF_PICK_FROM_SUBIMAGE method is never actually
signalled in the reference encoder, which is presumably why we haven't
spotted this before.

Change-Id: I745e2eab873c0b33920caca40e338af9d078d25e